tidied up a bit

+++ src/main/java/Calculator.java
@@ -12,7 +12,7 @@ public class Calculator {
         return a * b;
     }
 
-    public double divide(int a, int b){
-        return a/b;
+    public double divide(double a, double b){
+        return  a / b;
     }
 }
